# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Solennix API
  version: 1.0.0
  description: Event management platform for Latin American organizers
servers:
  - url: http://localhost:8000
    description: Local dev
  - url: https://api.solennix.com
    description: Production
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthSuccessResponse:
      type: object
      required:
        - user
        - tokens
      properties:
        user:
          $ref: "#/components/schemas/User"
        tokens:
          $ref: "#/components/schemas/TokenPair"
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 128
        name:
          type: string
          minLength: 1
          maxLength: 255
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code or type
        message:
          type: string
          description: Human-readable error message
...
paths:
...
  /api/auth/register:
    post:
      security: []
      operationId: register
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccessResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      security: []
      operationId: login
      summary: Authenticate with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccessResponse"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      security: []
      operationId: refreshToken
      summary: Refresh access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/auth/forgot-password:
    post:
      security: []
      operationId: forgotPassword
      summary: Start password reset flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset message accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /api/auth/reset-password:
    post:
      security: []
      operationId: resetPassword
      summary: Complete password reset with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or malformed request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      operationId: me
      summary: Get current authenticated user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "404":
          description: User not found
  /api/auth/change-password:
    post:
      operationId: changePassword
      summary: Change current user password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
  /api/auth/google:
    post:
      security: []
      operationId: googleSignIn
      summary: Sign in or register with Google ID token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleSignInRequest"
      responses:
        "200":
          description: Authenticated with Google
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccessResponse"
        "400":
          description: Invalid request body or missing id_token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid Google ID token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to link Google account or generate tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/apple:
    post:
      security: []
      operationId: appleSignIn
      summary: Sign in or register with Apple identity token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppleSignInRequest"
      responses:
        "200":
          description: Authenticated with Apple
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppleAuthSuccessResponse"
        "400":
          description: Invalid request body or missing identity_token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid Apple identity token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to link Apple account or generate tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/me:
    put:
      operationId: updateProfile
      summary: Update current user profile and business defaults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request body or contract template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "500":
          description: Failed to update profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/subscriptions/status:
    get:
      operationId: getSubscriptionStatus
      summary: Get current subscription and billing status
      responses:
        "200":
          description: Subscription status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubscriptionStatusResponse"
        "401":
          description: Unauthorized
  /api/subscriptions/checkout-session:
    post:
      operationId: createCheckoutSession
      summary: Create Stripe checkout session for Pro subscription
      responses:
        "200":
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "401":
          description: Unauthorized
        "500":
          description: Stripe not configured or session creation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/subscriptions/portal-session:
    post:
      operationId: createPortalSession
      summary: Create Stripe customer portal session
      responses:
        "200":
          description: Portal session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: No associated Stripe customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "500":
          description: Stripe not configured or portal creation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/subscriptions/webhook/stripe:
    post:
      security: []
      operationId: stripeWebhook
      summary: Receive Stripe webhook events
      responses:
        "200":
          description: Webhook processed
        "400":
          description: Invalid signature or payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Webhook not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/subscriptions/webhook/revenuecat:
    post:
      security: []
      operationId: revenueCatWebhook
      summary: Receive RevenueCat webhook events
      responses:
        "200":
          description: Webhook processed
        "400":
          description: Invalid webhook payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/events:
    get:
      operationId: listEvents
      summary: Get all events for authenticated user
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
        - name: start
          in: query
          schema:
            type: string
        - name: end
          in: query
          schema:
            type: string
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of events or paginated events when page is provided
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Event"
                  - $ref: "#/components/schemas/PaginatedEventsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createEvent
      summary: Create new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpsertRequest"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "422":
          description: Unprocessable entity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/events/{id}:
    get:
      operationId: getEvent
      summary: Get event by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
        "404":
          description: Event not found
    put:
      operationId: updateEvent
      summary: Update event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpsertRequest"
      responses:
        "200":
          description: Event updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "400":
          description: Validation error
        "401":
          description: Unauthorized
        "404":
          description: Event not found
    delete:
      operationId: deleteEvent
      summary: Delete event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Event deleted
        "401":
          description: Unauthorized
        "404":
          description: Event not found
  /api/events/search:
    get:
      operationId: searchEvents
      summary: Advanced search over events by query, status, date range or client
      description: At least one filter must be provided (q, status, from, to or client_id). Returns a flat array of events matching every provided filter.
      parameters:
        - name: q
          in: query
          description: Free-text query matched against title, description and notes
          schema:
            type: string
        - name: status
          in: query
          description: Filter by event status
          schema:
            type: string
            enum: [quoted, confirmed, completed, cancelled]
        - name: from
          in: query
          description: Inclusive start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Inclusive end date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: client_id
          in: query
          description: Restrict results to one client
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Matching events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "400":
          description: Invalid filters or missing mandatory filter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
        "500":
          description: Failed to search events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/events/upcoming:
    get:
      operationId: getUpcomingEvents
      summary: Get upcoming events for authenticated user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Upcoming events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
  /api/events/{id}/products:
    get:
      operationId: getEventProducts
      summary: Get products assigned to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Event products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventProduct"
        "400":
          description: Invalid event ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Event not found
...
  /api/dashboard/events-by-status:
    get:
      operationId: getEventsByStatus
      summary: Get event counts grouped by status
      responses:
        "200":
          description: Event counts by status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventStatusCount"
        "401":
          description: Unauthorized
...